Tangle `~/sbin/update-dms` script from niri config
This was an independent, erstwhile untracked, file for a hot minute, but
centralizing its definition with the niri one allows for easier
maintenance, cleaner git tracking of a platform-specific script, and
maybe also enables just running it inline with `C-c C-c`? Haven't tested
that last bit, but it's a clear win all around regardless.